fixed more clippy errors

+++ p12/src/main.rs
@@ -1,5 +1,5 @@
 fn main() {
     println!("Hello, world!");
     let a = 42u64;
-    let _ = a as u64;
+    let _ = a;
 }
